fixing chapters numbering
