+++ tests/books/toc/_toc_numbered_chapters.yml
@@ -1,0 +1,18 @@
+- file: index
+  title: Toc
+  numbered: true
+- chapter: Chapter 1
+  sections:
+  - file: content1
+    title: Content1
+  - file: subfolder/index
+    title: Subfolder
+    sections:
+    - file: subfolder/asubpage
+      title: Asubpage
+- chapter: Chapter 2
+  sections:
+  - file: content2
+    title: Content2
+  - file: content3
+    title: Content3
