+++ tests/books/toc/_toc_numbered_chapters_subset.yml
@@ -1,0 +1,21 @@
+- file: index
+  title: Toc
+- chapter: Chapter 1
+  numbered: true
+  sections:
+  - file: content1
+    title: Content1
+- chapter: Chapter 2
+  sections:
+  - file: content2
+    title: Content2
+  - file: content3
+    title: Content3
+- chapter: Chapter 3
+  numbered: true
+  sections:
+  - file: subfolder/index
+    title: Subfolder
+    sections:
+    - file: subfolder/asubpage
+      title: Asubpage
