+++ tests/test_toc.py
@@ -2,6 +2,10 @@ from pathlib import Path
 from subprocess import run, PIPE
 import pytest
 import yaml
+from bs4 import BeautifulSoup
+
+
+path_books = Path(__file__).parent.joinpath("books")
 
 
 def test_toc():
@@ -21,7 +25,7 @@ def test_toc():
 
 
 def test_toc_add_titles():
-    path_book = Path(__file__).parent.joinpath("books", "toc")
+    path_book = path_books.joinpath("toc")
     run(f"jb toc {path_book}".split(), check=True)
     toc_yaml = path_book.joinpath("_toc.yml")
     res = yaml.safe_load(toc_yaml.read_text(encoding="utf8"))
@@ -35,3 +39,30 @@ def test_toc_add_titles():
     assert "title" in res
     for section in res["sections"]:
         assert "title" in section
+
+
+def test_toc_numbered(tmpdir, file_regression):
+    """Testing that numbers make it into the sidebar"""
+    path_output = Path(tmpdir).joinpath("mybook").absolute()
+    toc_list = [
+        "_toc_numbered.yml",  # Numbered in top-level title
+        "_toc_numbered_chapters.yml",  # Numbered in top-level title w/ chapters
+        "_toc_numbered_chapters_subset.yml",  # Only some sections numbered
+    ]
+    for itoc in toc_list:
+        # Numbering with files
+        p_toc = path_books.joinpath("toc")
+        path_toc = p_toc.joinpath(itoc)
+        run(
+            f"jb build {p_toc} --path-output {path_output} --toc {path_toc} -W".split(),
+            check=True,
+        )
+
+        path_toc_directive = path_output.joinpath("_build", "html", "index.html")
+
+        # get the tableofcontents markup
+        soup = BeautifulSoup(
+            path_toc_directive.read_text(encoding="utf8"), "html.parser"
+        )
+        toc = soup.select("nav.bd-links")[0]
+        file_regression.check(str(toc), basename=itoc.split(".")[0], extension=".html")
