+++ tests/test_toc/_toc_numbered.html
@@ -1,0 +1,31 @@
+<nav aria-label="Main navigation" class="bd-links" id="bd-docs-nav">
+<ul class="nav sidenav_l1">
+<li class="toctree-l1">
+<a class="reference internal" href="#">
+   Main index
+  </a>
+</li>
+</ul>
+<ul class="nav sidenav_l1">
+<li class="toctree-l1">
+<a class="reference internal" href="content1.html">
+   1. Content1
+  </a>
+</li>
+<li class="toctree-l1">
+<a class="reference internal" href="subfolder/index.html">
+   2. Subfolder
+  </a>
+</li>
+<li class="toctree-l1">
+<a class="reference internal" href="content2.html">
+   3. Content2
+  </a>
+</li>
+<li class="toctree-l1">
+<a class="reference internal" href="content3.html">
+   4. Content3
+  </a>
+</li>
+</ul>
+</nav>
