+++ tests/test_toc/_toc_numbered_chapters.html
@@ -6,6 +6,11 @@
   </a>
 </li>
 </ul>
+<p class="caption">
+<span class="caption-text">
+  Chapter 1
+ </span>
+</p>
 <ul class="nav sidenav_l1">
 <li class="toctree-l1">
 <a class="reference internal" href="content1.html">
@@ -17,14 +22,21 @@
    2. Subfolder
   </a>
 </li>
+</ul>
+<p class="caption">
+<span class="caption-text">
+  Chapter 2
+ </span>
+</p>
+<ul class="nav sidenav_l1">
 <li class="toctree-l1">
 <a class="reference internal" href="content2.html">
-   3. Content2
+   1. Content2
   </a>
 </li>
 <li class="toctree-l1">
 <a class="reference internal" href="content3.html">
-   4. Content3
+   2. Content3
   </a>
 </li>
 </ul>
