+++ tests/test_toc/_toc_numbered_chapters_subset.html
@@ -1,0 +1,50 @@
+<nav aria-label="Main navigation" class="bd-links" id="bd-docs-nav">
+<ul class="nav sidenav_l1">
+<li class="toctree-l1">
+<a class="reference internal" href="#">
+   Main index
+  </a>
+</li>
+</ul>
+<p class="caption">
+<span class="caption-text">
+  Chapter 1
+ </span>
+</p>
+<ul class="nav sidenav_l1">
+<li class="toctree-l1">
+<a class="reference internal" href="content1.html">
+   1. Content1
+  </a>
+</li>
+</ul>
+<p class="caption">
+<span class="caption-text">
+  Chapter 2
+ </span>
+</p>
+<ul class="nav sidenav_l1">
+<li class="toctree-l1">
+<a class="reference internal" href="content2.html">
+   Content2
+  </a>
+</li>
+<li class="toctree-l1">
+<a class="reference internal" href="content3.html">
+   Content3
+  </a>
+</li>
+</ul>
+<p class="caption">
+<span class="caption-text">
+  Chapter 3
+ </span>
+</p>
+<ul class="nav sidenav_l1">
+<li class="toctree-l1">
+<a class="reference internal" href="subfolder/index.html">
+   1. Subfolder
+  </a>
+</li>
+</ul>
+</nav>
